rename file to match naming convention
